Break polygonFunctions into separate files
