+++ src/polygonFunctions/polygonAlloc.ts
@@ -1,0 +1,5 @@
+import { IPolygon } from "../types";
+
+export function polygonAlloc(): IPolygon {
+  return [];
+}
