+++ src/polygonFunctions/polygonGetNumSides.ts
@@ -1,0 +1,5 @@
+import { IPolygon } from "../types";
+
+export function polygonGetNumSides(poly: IPolygon) {
+  return poly.length / 2;
+}
