+++ src/polygonFunctions/polygonGetSideLength.ts
@@ -1,0 +1,7 @@
+import { _toPolyline } from "../internal/internalFunctions";
+import { polylineGetSegmentLength } from "../polylineFunctions/polylineGetSegmentLength";
+import { IPolygon } from "../types";
+
+export function polygonGetSideLength(poly: IPolygon, idx: number) {
+  return polylineGetSegmentLength(_toPolyline(poly), idx);
+}
