+++ src/polygonFunctions/polygonNearestVertexIndex.ts
@@ -1,0 +1,6 @@
+import { polylineNearestVertexIndex } from "../polylineFunctions/polylineNearestVertexIndex";
+import { IPolygon, IVec } from "../types";
+
+export function polygonNearestVertexIndex(poly: IPolygon, point: IVec) {
+  return polylineNearestVertexIndex(poly, point);
+}
